fix: fail to output json with batch option 'aws'
duck fails to output json report when AWS lambda fails.
The reason is that it cannot parse message including 'logUrl' to JSON.
This PR deletes 'logUrl' from error message to be able to parse it to JSON.

The output of FaastError.message has included logUrl since v3.0.14.
see the commit: https://github.com/faastjs/faast.js/commit/4c65cffa1e522e886f3948ee85b1943925e78281
So, with v3.0.13, this problem does not happen.

[FaastError document](https://faastjs.org/docs/api/faastjs.faasterror) indicates the below.
`The logUrl will be appended to the end of FaastError.message and FaastError.fullStack as well.`